+++ tokio/src/net/tcp/split_owned.rs
@@ -12,7 +12,6 @@ use crate::future::poll_fn;
 use crate::io::{AsyncRead, AsyncWrite, ReadBuf};
 use crate::net::TcpStream;
 
-use bytes::Buf;
 use std::error::Error;
 use std::net::Shutdown;
 use std::pin::Pin;
@@ -228,14 +227,6 @@ impl AsyncWrite for OwnedWriteHalf {
         buf: &[u8],
     ) -> Poll<io::Result<usize>> {
         self.inner.poll_write_priv(cx, buf)
-    }
-
-    fn poll_write_buf<B: Buf>(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut B,
-    ) -> Poll<io::Result<usize>> {
-        self.inner.poll_write_buf_priv(cx, buf)
     }
 
     #[inline]
